ml-model/requirements.txt
flask==3.0.3
flask-cors==4.0.1
tensorflow==2.16.1
pillow==10.3.0
numpy==1.26.4
python-dotenv==1.0.1
gunicorn==22.0.0
werkzeug==3.0.3
